fix(plan): split org-tree helpers into client-safe module + skip DB tests in CI
The browser was crashing on /plan/org-design Org Chart with
'Cannot read properties of undefined (reading Employee)' because
org-chart-view.tsx (a client component) imported buildOrgNeighborhood
from @ascenta/db/employees, which transitively loads the Mongoose
Employee schema. Mongoose can't run in a browser, so module evaluation
blew up.

Fix by extracting all the pure org-tree code (types, ROOT_SENTINELS,
buildOrgTree, buildOrgNeighborhood) into packages/db/src/org-tree.ts —
no Mongoose imports — and exposing it as @ascenta/db/org-tree. The
existing @ascenta/db/employees re-exports the same names for back-compat
with server-side callers (the API route, tests, etc.) so nothing else
moves. org-chart-view.tsx now imports from the safe path.

Also fix the related CI failure: api-org-tree.test.ts was hard-failing
on every CI run because MONGODB_URI isn't set there. Added the
describe.skipIf(SKIP_NO_DB) gate that the JD test files already use.

CLAUDE.md gains two pieces of guidance to prevent this from recurring:
expanded client-side import warning (pure helpers must live in their own
mongoose-free file) and a documented pattern for DB-backed tests.
