+++ packages/db/package.json
@@ -14,6 +14,7 @@
     "./documents": "./src/documents.ts",
     "./tracked-documents": "./src/tracked-documents.ts",
     "./employees": "./src/employees.ts",
+    "./org-tree": "./src/org-tree.ts",
     "./checkin-constants": "./src/checkin-constants.ts",
     "./checkin-schema": "./src/checkin-schema.ts",
     "./goal-constants": "./src/goal-constants.ts",
